+++ MatrixV3_View.h
@@ -5,5 +5,5 @@
 using namespace std;
 class MatrixV3_View : public Matrix {
 public:
-	MatrixV3_View(double posX, double posY, double posZ, double targetX, double targetY, double targetZ, double upX, double upY, double upZ);
+	MatrixV3_View(float posX, float posY, float posZ, float targetX, float targetY, float targetZ, float upX, float upY, float upZ);
 };
